refactor: fix filenaming towards consistent camelCase, step2
