+++ src/remoting/beanManager.js
